Fixed invalid/valid button in form

+++ santatecla-definiciones/src/main/java/com/group5/definitions/controllers/ConceptController.java
@@ -188,8 +188,7 @@ public class ConceptController {
 	}
 
 	@RequestMapping("/addAnswer/concept/{conceptId}")
-	public String addAnswer(Model model, @PathVariable String conceptId, 
-			/*@RequestParam String questionText,*/ //idk what this is for
+	public String addAnswer(Model model, @PathVariable String conceptId,
 			@RequestParam String justificationText, 
 			@RequestParam String answerText,
 			@RequestParam(value = "invalidJustification", required = false) String iJustification,
@@ -248,8 +247,14 @@ public class ConceptController {
 	@RequestMapping("/deleteJust/concept/{conceptId}/justification/{id}")
 	public void deleteJustification(Model model, @PathVariable String conceptId, @PathVariable long id,
 			HttpServletResponse httpServletResponse) throws IOException {
-		justificationService.deleteById(id);
-		httpServletResponse.sendRedirect("/concept/" + conceptId);
+		Answer ans = answerService.findByJustifications_Id(id);
+		System.out.println("prueba");
+		if((ans.getJustifications().size() == 1) && ans.isCorrect()) {
+			httpServletResponse.sendError(113, "prueba error");
+		}else {
+			justificationService.deleteById(id);
+			httpServletResponse.sendRedirect("/concept/" + conceptId);
+		}
 	}
 
 	@RequestMapping("/modifyJust/concept/{conceptId}/justificacion/{id}")
